quick move feedback queue under pr_iteration module
